+++ example/package.json
@@ -11,6 +11,8 @@
   },
   "dependencies": {
     "@pagopa/io-react-native-crypto": "^1.0.1",
+    "@react-navigation/native": "^7.0.14",
+    "@react-navigation/native-stack": "^7.2.0",
     "react": "18.3.1",
     "react-native": "0.76.7",
     "react-native-permissions": "^5.2.5",
